+++ types/openpgp/package.json
@@ -2,9 +2,9 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        ">=3.2.0-0": {
+        "<=3.1": {
             "*": [
-                "ts3.2/*"
+                "ts3.1/*"
             ]
         }
     }
